requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.62.0
Flask>=2.3.0
Werkzeug>=2.3.0
opencv-python>=4.7.0
scipy>=1.9.0
scikit-image>=0.19.0